make authorization

+++ app/Http/Controllers/Dashboard/OfferController.php
@@ -11,12 +11,16 @@ class OfferController extends Controller
     //
     public function index()
     {
+        $this->authorize('viewAny', Offer::class);
+
         $offers = Offer::with('images', 'product')->get();
         return view('offers.index', compact('offers'));
     }
 
     public function destroy(Offer $offer)
     {
+        $this->authorize('delete', $offer);
+
         $isDeleted = $offer->delete();
         return response()->json([
             'message' => $isDeleted ? 'تم حذف العرض بنجاح' : 'فشل حذف العرض',
